% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMatMR-package.R
\name{fmm_to_mat}
\alias{fmm_to_mat}
\title{Convert Matrix Market File to Matrix}
\arguments{
\item{filename}{The name of the input Matrix Market file to be read.}
}
\value{
A matrix containing the data read from the Matrix Market file.
}
\description{
This function reads a Matrix Market file and converts it to a matrix in R.
}
\examples{
\dontrun{
mat <- fmm_to_mat("matrix.mtx")
}
}
