% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMatMR-package.R
\name{vec_to_fmm}
\alias{vec_to_fmm}
\title{Convert a Numeric Vector to Matrix Market Format}
\arguments{
\item{input}{A numeric vector to be converted.}

\item{filename}{The name of the output file where the Matrix Market formatted
data will be saved.}
}
\value{
A boolean indicating success or failure. Writes a MTX file to disk.
}
\description{
This function takes a numeric vector and converts it into a
Matrix Market output file.
}
\examples{
vec <- c(1, 2, 3)
vec_to_fmm(vec, tempfile(fileext = ".mtx"))
}
